feat: Implement FIDO CTAP2 protocol for credential and PIN management.

+++ firmware/src/usb_descriptors.c
@@ -130,7 +130,7 @@ char const *string_desc_arr[] = {
 
 static uint16_t _desc_str[32];
 
-uint8_t const *tud_descriptor_string_cb(uint8_t index, uint16_t langid) {
+uint16_t const *tud_descriptor_string_cb(uint8_t index, uint16_t langid) {
   (void)langid;
   uint8_t chr_count;
